requirements.txt
absl-py==1.4.0
aiohttp==3.8.5
aiosignal==1.3.1
alabaster==0.7.16
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.3
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
cachelib==0.10.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colour==0.1.5
comm==0.1.3
contourpy==1.1.0
cycler==0.11.0
dataclasses-json==0.6.4
debugpy==1.6.7
decorator==5.1.1
distlib==0.3.7
docopt==0.6.2
docutils==0.20.1
executing==1.2.0
filelock==3.12.2
fire==0.5.0
Flask==2.3.3
Flask-OAuthlib==0.9.6
flatbuffers==23.5.26
fonttools==4.40.0
forex-python==1.8
frozenlist==1.4.0
fsspec==2023.6.0
gast==0.4.0
geojson==2.5.0
glob2==0.7
google-ai-generativelanguage==0.4.0
google-api-core==2.11.1
google-api-python-client==2.110.0
google-auth==2.22.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-generativeai==0.3.2
google-pasta==0.2.0
googleapis-common-protos==1.60.0
grpcio==1.60.0
grpcio-status==1.60.0
gspread==5.12.4
gunicorn==21.2.0
h11==0.14.0
h5py==3.9.0
httpcore==1.0.4
httplib2==0.22.0
httpx==0.25.2
huggingface-hub==0.16.4
idna==3.4
imagesize==1.4.1
indic-nlp-library==0.92
ipykernel==6.24.0
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.3.0
jupyter_core==5.3.1
keras==2.13.1
kiwisolver==1.4.4
langchain==0.1.11
langchain-community==0.0.27
langchain-core==0.1.30
langchain-text-splitters==0.0.1
langsmith==0.1.23
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.3
marshmallow==3.21.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
Morfessor==2.0.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.1
numpy==1.24.3
oauth2client==4.1.3
oauthlib==3.2.2
ollama==0.1.6
openai==0.28.0
opencv-python==4.8.0.74
opt-einsum==3.3.0
orjson==3.9.15
packaging==23.2
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pipreqs==0.4.13
platformdirs==3.10.0
prompt-toolkit==3.0.39
proto-plus==1.23.0
protobuf==4.23.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.6.3
pydantic_core==2.16.3
Pygments==2.15.1
pynput==1.7.6
pyobjc-core==10.1
pyobjc-framework-ApplicationServices==10.1
pyobjc-framework-Cocoa==10.1
pyobjc-framework-CoreText==10.1
pyobjc-framework-Quartz==10.1
pyowm==3.3.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.0
redis==5.0.1
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.3.2
scikit-learn==1.3.0
scipy==1.11.1
sentencepiece==0.1.99
simplejson==3.19.2
six==1.16.0
sklearn==0.0.post7
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.5
Sphinx==7.2.6
sphinx-argparse==0.4.0
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
splitwise==3.0.0
spotipy==2.23.0
SQLAlchemy==2.0.28
stack-data==0.6.2
stripe==7.7.0
sympy==1.12
tenacity==8.2.3
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-macos==2.13.0
tensorflow-metal==1.0.1
termcolor==2.3.0
threadpoolctl==3.1.0
tk==0.1.0
tokenizers==0.13.3
torch==2.0.1
tornado==6.3.2
tqdm==4.66.1
traitlets==5.9.0
transformers==4.31.0
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
virtualenv==20.24.5
wcwidth==0.2.6
Werkzeug==2.3.7
wrapt==1.15.0
yarg==0.1.9
yarl==1.9.2
youtube-transcript-api==0.6.1
